+++ lib/src/framework/framework.dart
@@ -1,0 +1,13 @@
+import 'dart:collection';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/scheduler.dart';
+import 'package:flutter/widgets.dart';
+
+import '../../provide_it.dart';
+import '../utils/tree_map.dart';
+
+part 'provide_it_root.dart';
+part 'ref.dart';
+part 'ref_state.dart';
+part 'ref_widget.dart';