+++ lib/src/framework/provide_it_root.dart
@@ -1,13 +1,7 @@
-part of '../framework.dart';
+part of 'framework.dart';
 
 class ProvideItRoot extends InheritedWidget {
   const ProvideItRoot({super.key, required super.child});
-
-  static ProvideItRootElement of(BuildContext context) {
-    final el = context.getElementForInheritedWidgetOfExactType<ProvideItRoot>();
-    assert(el != null, 'You must set `ProvideIt.root` in your app.');
-    return el! as ProvideItRootElement;
-  }
 
   @override
   InheritedElement createElement() => ProvideItRootElement(this);
@@ -17,27 +11,43 @@ class ProvideItRoot extends InheritedWidget {
 }
 
 class ProvideItRootElement extends InheritedElement {
-  ProvideItRootElement(super.widget);
+  ProvideItRootElement._(super.widget);
   bool _reassembled = false;
   bool _doingInit = false;
 
+  factory ProvideItRootElement(ProvideItRoot widget) {
+    assert(_instance == null, 'You can only have one `ProvideIt.root`.');
+    return _instance ??= ProvideItRootElement._(widget);
+  }
+
+  static ProvideItRootElement? _instance;
+  static ProvideItRootElement get instance {
+    assert(_instance != null, 'You must set `ProvideIt.root` in your app.');
+    return _instance!;
+  }
+
   bool get debugDoingInit => _doingInit;
 
-  // state tree
+  // state binder tree by context and index.
   final _tree = TreeMap<Element, TreeMap<int, _State>>();
-  final _index = <Element, int>{};
+  final _treeIndex = <Element, int>{};
 
-  // state cache
+  // state finder cache by context, type and key.
   final _cache = <BuildContext, Map<Type, Map<Object?, _State?>>>{};
+  final _cacheIndex = <BuildContext, int>{};
+
+  void _assert(BuildContext context, String method) {
+    assert(
+      context is Element && context.debugDoingBuild,
+      '$method() should be called within the build() method of a widget.',
+    );
+  }
 
   int _initIndex(BuildContext context) {
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    SchedulerBinding.instance.addPostFrameCallback((_) {
       // we reset the index for the next build.
-      final index = _index.remove(context) ?? 0;
-      final length = _tree[context]?.length ?? 0;
-
-      // if a ref was removed, we dispose it.
-      if (index < length) _disposeIndex(context, index);
+      _treeIndex.remove(context);
+      _cacheIndex.remove(context);
       _reassembled = false;
     });
 
@@ -45,18 +55,12 @@ class ProvideItRootElement extends InheritedElement {
   }
 
   _State<T> _state<T>(BuildContext context, Ref<T> ref) {
-    assert(context is Element);
-    assert(
-      context.debugDoingBuild,
-      'build() should be called within the build method of a widget.',
-    );
-
     // we depend so we can get notified by [removeDependent].
     context.dependOnInheritedElement(this);
 
     final branch = _tree[context as Element] ??= TreeMap<int, _State>();
-    final index = _index[context] ??= _initIndex(context);
-    _index[context] = index + 1;
+    final index = _treeIndex[context] ??= _initIndex(context);
+    _treeIndex[context] = index + 1;
 
     _State<T> create() {
       _doingInit = true;
@@ -78,8 +82,6 @@ class ProvideItRootElement extends InheritedElement {
     }
 
     _State<T> reset(_State<dynamic> old) {
-      _disposeIndex(context, index);
-
       if (_reassembled) return create();
       throw StateError('${old.ref.runtimeType} != ${ref.runtimeType}');
     }
@@ -92,11 +94,21 @@ class ProvideItRootElement extends InheritedElement {
   }
 
   _State<T> _stateOf<T>(BuildContext context, {Object? key}) {
+    // we depend so we can get notified by [removeDependent].
+    context.dependOnInheritedElement(this);
+
     final contextCache = _cache[context] ??= {};
     final typeCache = contextCache[T] ??= HashMap(equals: Ref.equals);
     final state = typeCache[key] ??= _findState<T>(key: key);
 
-    if (state case _State<T> state) return state;
+    if (state case _State<T> state) {
+      final index = _cacheIndex[context] ??= _initIndex(context as Element);
+      _cacheIndex[context] = index + 1;
+
+      return state;
+    }
+
+    // we assume it's a global lazy ref, so we can bind/read it.
     if (key case Ref<T> ref) return _state(context, ref);
 
     throw StateError('No state found for $T with key $key');
@@ -114,23 +126,29 @@ class ProvideItRootElement extends InheritedElement {
   }
 
   R bind<R, T>(BuildContext context, Ref<T> ref) {
+    _assert(context, 'bind');
+
     return _state(context, ref).build(context) as R;
   }
 
-  T read<T>(BuildContext context, {Object? key}) {
-    return _stateOf<T>(context, key: key).read(context);
-  }
-
   T watch<T>(BuildContext context, {Object? key}) {
+    _assert(context, 'watch');
+
     return _stateOf<T>(context, key: key).watch(context);
   }
 
   R select<T, R>(BuildContext context, R selector(T value), {Object? key}) {
-    return _stateOf<T>(context, key: key).select(context, selector);
+    _assert(context, 'select');
+
+    final state = _stateOf<T>(context, key: key);
+    return state.select(context, _cacheIndex[context]!, selector);
   }
 
   void listen<T>(BuildContext context, void listener(T value), {Object? key}) {
-    _stateOf<T>(context, key: key).listen(context, listener);
+    _assert(context, 'listen');
+
+    final state = _stateOf<T>(context, key: key);
+    state.listen(context, _cacheIndex[context]!, listener);
   }
 
   void listenSelect<T, R>(
@@ -139,18 +157,29 @@ class ProvideItRootElement extends InheritedElement {
     void listener(R previous, R next), {
     Object? key,
   }) {
-    _stateOf<T>(context, key: key).listenSelect(context, selector, listener);
+    _assert(context, 'listenSelect');
+
+    final state = _stateOf<T>(context, key: key);
+    state.listenSelect(context, _cacheIndex[context]!, selector, listener);
   }
+
+  T read<T>(BuildContext context, {Object? key}) {
+    final state = _stateOf<T>(context, key: key);
+
+    return state._lastReadValue = state.read(context);
+  }
+
+  T readIt<T>({Object? key}) => read<T>(this, key: key);
 
   @override
   void reassemble() {
-    super.reassemble();
-    _reassembled = true;
     for (final branch in _tree.values) {
       for (final state in branch.values) {
         state.reassemble();
       }
     }
+    super.reassemble();
+    _reassembled = true;
   }
 
   @override
@@ -166,29 +195,38 @@ class ProvideItRootElement extends InheritedElement {
     WidgetsBinding.instance.addPostFrameCallback((_) {
       // we need to check if the dependent is still mounted
       // because it could have been removed from the tree.
-      // if it's still mounted, we reactivate the providers.
+      // if it's still mounted, we reactivate.
       dependent.mounted
-          ? _tree[dependent]?.forEach((_, it) => it.activate())
-          : _disposeDependents(dependent);
+          ? _tree[dependent]?.forEach((_, state) => state.activate())
+          : _disposeDependent(dependent);
     });
     super.removeDependent(dependent);
   }
 
-  /// Called when any [Ref] is removed from this [context].
-  void _disposeIndex(BuildContext context, int index) {
-    final branch = _tree[context] ?? TreeMap<int, _State>();
-    final indexes = branch.keys.where((i) => i >= index);
+  /// Called when a [context] is unmounted from the tree.
+  void _disposeDependent(Element context) {
+    // we dispose all bindings of this context.
+    _tree.remove(context)?.forEach((_, state) => state.dispose());
+    _treeIndex.remove(context);
+    _cacheIndex.remove(context);
 
-    // we dispose current index and subsequents as they are displaced.
-    for (var index in indexes) {
-      branch.remove(index)?.dispose();
+    // we remove all dependencies of this context.
+    _cache.remove(context)?.forEach((_, branch) =>
+        branch.forEach((_, state) => state?.removeDependent(context)));
+  }
+
+  /// Called when a [ref] self-disposes on reassemble.
+  void _disposeRef<T>(BuildContext context, Ref<T> ref) {
+    final branch = _tree[context] ?? TreeMap<int, _State>();
+
+    for (var e in branch.entries.toList()) {
+      if (e.value.ref == ref) return branch.remove(e.key)?.dispose();
     }
   }
 
-  /// Called when a [context] is removed from the tree.
-  void _disposeDependents(BuildContext context) {
-    _tree.remove(context)?.forEach((_, state) => state.dispose());
-    _index.remove(context);
-    _cache.remove(context);
+  void debugTree() {
+    if (kDebugMode) {
+      print(_tree);
+    }
   }
 }