+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.1.0
+
+- Changed `Provider` to `ProvideIt.root`.
+- Improved ref dispose/removal on hot-reload.
+- Improved several `toString` for debugging.
+- Added several `asserts` for debugging.
+- Added `readIt` locator.
+- Added `Ref.debugValue` for debugging.
+- Added `ProvideIt.log` for debugging.
+- Added `RefState.removeDependent`.
+- Added `dependOnInheritedElement` for bind accessors.
+- Removed the need for `RefState._garbageCollector`.
+
 ## 0.0.1
 
 - Initial release.